requirements.txt
elasticsearch==6.0.0
elasticsearch-dsl==6.0.1
python-dateutil==2.6.1
python-geoip-geolite2==2015.303
six==1.11.0
urllib3==1.22
